+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 - Change dependency to version `0.4.0` of other arkwork-rs crates.
 - Fiat-Shamir transformation for the AHP uses the Poseidon sponge function.
-- Introduced Absorb trait bound on field elements to facilitate more efficient absorption methods.
+- Introduced fast_prove and fast_verify methods for PrimeFields that also implement the Absorb trait.
 - Added RngCore and CryptographicSponge traits for rng sources.
 
 ## v0.3.0